Improved documentation of background recovery command timer. Moved to AsyncConcurrentQueue to utils.
